React routing implemented

+++ src/containers/LessonTabs.js
@@ -19,6 +19,7 @@ export default class LessonTabs
         this.createLesson = this.createLesson.bind(this);
         this.deleteLesson = this.deleteLesson.bind(this);
         this.titleChanged = this.titleChanged.bind(this);
+       // this.displayTopics = this.displayTopics.bind(this);
 
         this.lessonService = LessonService.instance;
     }
@@ -28,6 +29,7 @@ export default class LessonTabs
         this.setModuleId(this.props.moduleId);
         console.log(this.props.courseId);
         console.log(this.props.moduleId);
+        this.findAllLessonsForModule(this.props.courseId, this.props.moduleId)
     }
 
     componentWillReceiveProps(newProps){
@@ -65,52 +67,82 @@ export default class LessonTabs
             .then(() => { this.findAllLessonsForModule(this.props.courseId, this.props.moduleId); });
     }
 
+    // displayTopics(lessonId) {
+    //     this.props.displayTopics(lessonId);
+    // }
+
     deleteLesson(lessonId) {
+        var answer = window.confirm("Click Ok to Delete");
+        if (answer == true) {
+            this.lessonService
+                .deleteLesson(lessonId)
+                .then(() => {
+                        this.findAllLessonsForModule
+                        (this.state.courseId, this.state.moduleId)
+                    }
+                );
+        }
+    }
+
+    findAllLessons() {
         this.lessonService
-            .deleteLesson(lessonId)
-            .then(() => {
-                    this.findAllLessonsForModule
-                    (this.state.courseId, this.state.moduleId)
-                }
-            );
+            .findAllLessons()
+            .then((lessons) => {
+                this.setState({lessons: lessons});
+            })
     }
 
     renderListOfLessons() {
         let lessons = null;
         if(this.state) {
-            lessons = this.state.lessons.map(
-                function (lesson) {
+            lessons = this.state.lessons.map((lesson) => {
                     return <LessonListItem key={lesson.id}
                                            lesson={lesson}/>
                 }
-            )
-            lessons = this.state.lessons.map((lesson) => {
-                return <LessonListItem lesson={lesson}
-                                       key={lesson.id}
-                                       deleteLesson={this.deleteLesson}/>
-            });
+            );
         }
         return (
             lessons
         )
     }
 
+    // render() {
+    //     return (
+    //         <div className="container-fluid">
+    //             <h4>Lesson List for module: {this.state.moduleId}</h4>
+    //             <ul className="nav nav-pills">
+    //                 {this.renderListOfLessons()}
+    //             </ul>
+    //                 <input onChange={this.titleChanged}
+    //                        value={this.state.lesson.title}
+    //                        placeholder="add lesson"
+    //                        className="font-weight-bold text-center"
+    //                        style = {{marginLeft: 20, marginTop: 20}}/>
+    //                 <button onClick={this.createLesson} className="btn btn-dark">
+    //                       <i className="fa fa-plus"></i>
+    //                 </button>
+    //         </div>
+    //     );
+    // }
+
     render() {
         return (
-            <div className="container-fluid">
-                <h4>Lesson List for module: {this.state.moduleId}</h4>
+            <div>
+                <h3 style={{textAlign: "left"}}>Lessons</h3>
+                <br/>
                 <ul className="nav nav-pills">
                     {this.renderListOfLessons()}
                 </ul>
-                    <input onChange={this.titleChanged}
-                           value={this.state.lesson.title}
-                           placeholder="add lesson"
-                           className="font-weight-bold text-center"
-                           style = {{marginLeft: 20, marginTop: 20}}/>
-                    <button onClick={this.createLesson} className="btn btn-dark">
-                          <i className="fa fa-plus"></i>
-                    </button>
+                <br/>
+                <input onChange={this.titleChanged}
+                       value={this.state.lesson.title}
+                       placeholder="Add Lesson"
+                       className="form-control text-center font-weight-bold"/>
+                <button onClick={this.createLesson} className="btn btn-primary btn-block">
+                    <i className="fa fa-plus"></i>
+                </button>
             </div>
         );
     }
+
 }
